# openapi.json
{
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/categories": {
			"post": {
				"operationId": "Categories_create",
				"parameters": [ ],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCategoryDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryDto"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			},
			"get": {
				"operationId": "Categories_findAll",
				"parameters": [
					{
						"name": "depth",
						"required": false,
						"in": "query",
						"description": "The depth of children to retrieve. Set 0 for only the root categories, -1 for flat array, leave empty for the full tree",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/CategoryDto"
									}
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/categories/slugs/{slug}": {
			"get": {
				"operationId": "Categories_findOneBySlug",
				"summary": "Get a category by slug",
				"description": "Gets a category by slug. Use depth query to control the level of children to retrieve",
				"parameters": [
					{
						"name": "slug",
						"required": true,
						"in": "path",
						"description": "The slug of the category to get",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "depth",
						"required": false,
						"in": "query",
						"description": "The depth of children to retrieve. Set 0 for only the requested category, leave empty for the full tree",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryDto"
								}
							}
						}
					},
					"404": {
						"description": "The category does not exists"
					}
				},
				"tags": [
					"Categories"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/categories/ids": {
			"get": {
				"operationId": "Categories_getAllIds",
				"summary": "",
				"description": "Gets all ids of categories",
				"parameters": [ ],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"type": "number"
									}
								}
							}
						}
					},
					"404": {
						"description": "The category does not exists"
					}
				},
				"tags": [
					"Categories"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/categories/slugs": {
			"get": {
				"operationId": "Categories_getAllSlugs",
				"summary": "",
				"description": "Gets all slugs of categories",
				"parameters": [ ],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"type": "string"
									}
								}
							}
						}
					},
					"404": {
						"description": "The category does not exists"
					}
				},
				"tags": [
					"Categories"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/categories/{id}": {
			"get": {
				"operationId": "Categories_findOne",
				"summary": "Get a category by id",
				"description": "Gets a category. Use depth query to control the level of children to retrieve",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "The id of the category to get",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "depth",
						"required": false,
						"in": "query",
						"description": "The depth of children to retrieve. Set 0 for only the requested category, leave empty for the full tree",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryDto"
								}
							}
						}
					},
					"404": {
						"description": "The category does not exists"
					}
				},
				"tags": [
					"Categories"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			},
			"put": {
				"operationId": "Categories_update",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCategoryDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryDto"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			},
			"delete": {
				"operationId": "Categories_remove",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "forceDelete",
						"required": false,
						"in": "query",
						"description": "Deletes the children subcategories. Default to false.",
						"schema": {
							"type": "boolean"
						}
					}
				],
				"responses": {
					"200": {
						"description": ""
					}
				},
				"tags": [
					"Categories"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/categories/{id}/ancestors": {
			"get": {
				"operationId": "Categories_findAncestorsById",
				"summary": "Get ancestors of category by id",
				"description": "Gets all the ancestors of a category. Useful for breadcrumbs",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "The id of the category to get",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryDto"
								}
							}
						}
					},
					"404": {
						"description": "The category does not exists"
					}
				},
				"tags": [
					"Categories"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
...
		"/api/v1/discounts/{id}/categories/{categoryId}": {
			"delete": {
				"operationId": "DiscountSales_removeCategory",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "categoryId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DiscountSaleDto"
								}
							}
						}
					}
				},
				"tags": [
					"Discount"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
...
		"/api/v1/address/{id}": {
			"get": {
				"operationId": "Address_getById",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AddressDto"
								}
							}
						}
					}
				},
				"tags": [
					"Address"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			},
			"put": {
				"operationId": "Address_updateById",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAddressDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AddressDto"
								}
							}
						}
					}
				},
				"tags": [
					"Address"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
...
		"/api/v1/checkout": {
			"post": {
				"operationId": "Checkout_create",
				"parameters": [ ],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCheckoutDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CheckoutDto"
								}
							}
						}
					}
				},
				"tags": [
					"Checkout"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/checkout/{token}": {
			"get": {
				"operationId": "Checkout_findOne",
				"parameters": [
					{
						"name": "token",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CheckoutDto"
								}
							}
						}
					}
				},
				"tags": [
					"Checkout"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			},
			"patch": {
				"operationId": "Checkout_update",
				"parameters": [
					{
						"name": "token",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCheckoutDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CheckoutDto"
								}
							}
						}
					}
				},
				"tags": [
					"Checkout"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/checkout/{token}/lines": {
			"patch": {
				"operationId": "Checkout_updateLines",
				"parameters": [
					{
						"name": "token",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCheckoutLineDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CheckoutDto"
								}
							}
						}
					}
				},
				"tags": [
					"Checkout"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/checkout/{token}/address": {
			"patch": {
				"operationId": "Checkout_updateAddress",
				"parameters": [
					{
						"name": "token",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "shipping",
						"required": false,
						"in": "query",
						"schema": {
							"type": "boolean"
						}
					},
					{
						"name": "billing",
						"required": false,
						"in": "query",
						"schema": {
							"type": "boolean"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAddressDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CheckoutDto"
								}
							}
						}
					}
				},
				"tags": [
					"Checkout"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/checkout/{token}/voucher": {
			"patch": {
				"operationId": "Checkout_updateVoucher",
				"parameters": [
					{
						"name": "token",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCheckoutVoucherDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CheckoutDto"
								}
							}
						}
					}
				},
				"tags": [
					"Checkout"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
...
	},
	"info": {
		"title": "Nima CMS",
		"description": "The core api for the Nima CMS backend.",
		"version": "0.0.1-alpha",
		"contact": {}
	},
...
	"servers": [
		{
			"url": "http://localhost:3333",
			"description": "Localhost"
		}
	],
	"components": {
		"securitySchemes": {
			"bearer": {
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"type": "http"
			}
		},
		"schemas": {
			"TranslatableDto": {
				"type": "object",
				"properties": {
					"de": {
						"type": "string"
					},
					"it": {
						"type": "string"
					},
					"fr": {
						"type": "string"
					},
					"es": {
						"type": "string"
					},
					"el": {
						"type": "string"
					},
					"en": {
						"type": "string"
					}
				}
			},
...
			"LanguageCode": {
				"type": "string",
				"enum": [
					"en",
					"el",
					"fr",
					"es",
					"it",
					"de",
					"ru"
				]
			},
...
			"CreateCategoryDto": {
				"type": "object",
				"properties": {
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"seoTitle": {
						"example": {
							"en": "Category Seo Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"seoDescription": {
						"example": {
							"en": "Category Seo Description"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"name": {
						"example": {
							"en": "Category Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"description": {
						"example": {
							"en": "Category Description"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"slug": {
						"type": "string",
						"example": "category-name"
					},
					"parentId": {
						"type": "number",
						"example": 1
					}
				},
				"required": [
					"privateMetadata",
					"metadata",
					"seoTitle",
					"seoDescription",
					"name",
					"description",
					"slug"
				]
			},
			"CategoryDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number",
						"example": 1
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"seoTitle": {
						"example": {
							"en": "Category Seo Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"seoDescription": {
						"example": {
							"en": "Category Seo Description"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"name": {
						"example": {
							"en": "Category Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"description": {
						"example": {
							"en": "Category Description"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"slug": {
						"type": "string",
						"example": "category-name"
					},
					"children": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CategoryDto"
						}
					},
					"parent": {
						"description": "Only for ancestors query",
						"allOf": [
							{
								"$ref": "#/components/schemas/CategoryDto"
							}
						]
					}
				},
				"required": [
					"id",
					"privateMetadata",
					"metadata",
					"seoTitle",
					"seoDescription",
					"name",
					"description",
					"slug",
					"children"
				]
			},
			"UpdateCategoryDto": {
				"type": "object",
				"properties": {
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"seoTitle": {
						"example": {
							"en": "Category Seo Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"seoDescription": {
						"example": {
							"en": "Category Seo Description"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"name": {
						"example": {
							"en": "Category Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"description": {
						"example": {
							"en": "Category Description"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"slug": {
						"type": "string",
						"example": "category-name"
					},
					"parentId": {
						"type": "number",
						"example": 1
					}
				},
				"required": [
					"privateMetadata",
					"metadata",
					"seoTitle",
					"seoDescription",
					"name",
					"description",
					"slug"
				]
			},
			"DiscountType": {
				"type": "string",
				"enum": [
					"PERCENTAGE",
					"FLAT"
				]
			},
...
			"DiscountSaleDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number",
						"example": 1
					},
					"name": {
						"example": {
							"en": "Discount Sale Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"discountType": {
						"$ref": "#/components/schemas/DiscountType"
					},
					"discountValue": {
						"type": "number"
					},
					"endDate": {
						"type": "string",
						"example": "2022-12-31"
					},
					"startDate": {
						"type": "string",
						"example": "2022-01-01"
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"created": {
						"type": "string",
						"example": "2022-01-01"
					},
					"updatedAt": {
						"type": "string",
						"example": "2022-01-01"
					},
					"categories": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CategoryDto"
						}
					},
					"products": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ProductDto"
						}
					},
					"variants": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ProductVariantDto"
						}
					},
					"collections": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CollectionDto"
						}
					}
				},
				"required": [
					"id",
					"name",
					"discountType",
					"discountValue",
					"metadata",
					"privateMetadata",
					"created",
					"updatedAt",
					"categories",
					"products",
					"variants",
					"collections"
				]
			},
...
			"AddressDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number",
						"example": 1
					},
					"firstName": {
						"type": "string",
						"example": "John"
					},
					"lastName": {
						"type": "string",
						"example": "Doe"
					},
					"companyName": {
						"type": "string",
						"example": "Indera Inc."
					},
					"phone": {
						"type": "string",
						"example": "+1 1234 567 890"
					},
					"country": {
						"type": "string",
						"example": "Greece"
					},
					"state": {
						"type": "string",
						"example": "Attica"
					},
					"city": {
						"type": "string",
						"example": "Athens"
					},
					"zip": {
						"type": "string",
						"example": "12345"
					},
					"address": {
						"type": "string",
						"example": "Example 123"
					},
					"address2": {
						"type": "string",
						"example": "Example 123"
					}
				},
				"required": [
					"id",
					"country",
					"state",
					"city",
					"zip",
					"address"
				]
			},
...
			"CreateAddressDto": {
				"type": "object",
				"properties": {
					"firstName": {
						"type": "string",
						"example": "John"
					},
					"lastName": {
						"type": "string",
						"example": "Doe"
					},
					"companyName": {
						"type": "string",
						"example": "Indera Inc."
					},
					"phone": {
						"type": "string",
						"example": "+1 1234 567 890"
					},
					"country": {
						"type": "string",
						"example": "Greece"
					},
					"state": {
						"type": "string",
						"example": "Attica"
					},
					"city": {
						"type": "string",
						"example": "Athens"
					},
					"zip": {
						"type": "string",
						"example": "12345"
					},
					"address": {
						"type": "string",
						"example": "Example 123"
					},
					"address2": {
						"type": "string",
						"example": "Example 123"
					}
				},
				"required": [
					"country",
					"state",
					"city",
					"zip",
					"address"
				]
			},
...
			"PaymentMethod": {
				"type": "string",
				"enum": [
					"CARD",
					"CASH_ON_DELIVERY",
					"BANK_TRANSFER"
				]
			},
...
			"CreateCheckoutDto": {
				"type": "object",
				"properties": {
					"languageCode": {
						"$ref": "#/components/schemas/LanguageCode"
					}
				},
				"required": [
					"languageCode"
				]
			},
...
			"CheckoutLineDto": {
				"type": "object",
				"properties": {
					"quantity": {
						"type": "number",
						"example": 1
					},
					"variantId": {
						"type": "number"
					},
					"productId": {
						"type": "number"
					},
					"totalCost": {
						"type": "number"
					},
					"discountedTotalCost": {
						"type": "number"
					}
				},
				"required": [
					"quantity",
					"variantId",
					"productId",
					"totalCost",
					"discountedTotalCost"
				]
			},
...
			"CheckoutDto": {
				"type": "object",
				"properties": {
					"created": {
						"type": "string",
						"example": "2022-01-01"
					},
					"lastChange": {
						"type": "string",
						"example": "2022-01-01"
					},
					"email": {
						"type": "string",
						"example": "test@example.com"
					},
					"useShippingAsBilling": {
						"type": "boolean",
						"example": true
					},
					"token": {
						"type": "string",
						"example": ""
					},
					"user": {
						"$ref": "#/components/schemas/UserDto"
					},
					"billingAddress": {
						"$ref": "#/components/schemas/AddressDto"
					},
					"discountAmount": {
						"type": "number",
						"example": 12.3
					},
					"discountName": {
						"type": "string",
						"example": "discount-name"
					},
					"note": {
						"type": "string",
						"example": "checkout-note"
					},
					"shippingAddress": {
						"$ref": "#/components/schemas/AddressDto"
					},
					"voucherCode": {
						"type": "string",
						"example": "ASDF123"
					},
					"translatedDiscountName": {
						"type": "string",
						"example": "discount-name"
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"currency": {
						"type": "string",
						"example": "EUR"
					},
					"redirectUrl": {
						"type": "string",
						"example": "http://www.example.com/redirect"
					},
					"trackingCode": {
						"type": "string",
						"example": "ASDF123"
					},
					"languageCode": {
						"$ref": "#/components/schemas/LanguageCode"
					},
					"paymentMethod": {
						"$ref": "#/components/schemas/PaymentMethod"
					},
					"lines": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CheckoutLineDto"
						}
					},
					"subtotalPrice": {
						"type": "number"
					},
					"shippingCost": {
						"type": "number"
					},
					"quantity": {
						"type": "number"
					},
					"discount": {
						"type": "number"
					},
					"totalCost": {
						"type": "number"
					},
					"shippingMethod": {
						"$ref": "#/components/schemas/ShippingMethodDto"
					},
					"availableShippingMethods": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CheckoutAvailableShippingDto"
						}
					}
				},
				"required": [
					"created",
					"lastChange",
					"token",
					"billingAddress",
					"discountAmount",
					"note",
					"shippingAddress",
					"metadata",
					"privateMetadata",
					"currency",
					"languageCode",
					"paymentMethod",
					"lines",
					"subtotalPrice",
					"shippingCost",
					"quantity",
					"discount",
					"totalCost",
					"availableShippingMethods"
				]
			},
			"UpdateCheckoutDto": {
				"type": "object",
				"properties": {
					"email": {
						"type": "string",
						"example": "test@example.com"
					},
					"useShippingAsBilling": {
						"type": "boolean",
						"example": true
					},
					"note": {
						"type": "string",
						"example": "checkout-note"
					},
					"languageCode": {
						"$ref": "#/components/schemas/LanguageCode"
					},
					"paymentMethod": {
						"$ref": "#/components/schemas/PaymentMethod"
					},
					"shippingMethodId": {
						"type": "number"
					}
				}
			},
			"UpdateCheckoutLineDto": {
				"type": "object",
				"properties": {
					"quantity": {
						"type": "number",
						"example": 1
					},
					"variantId": {
						"type": "number"
					}
				},
				"required": [
					"quantity",
					"variantId"
				]
			},
			"UpdateCheckoutVoucherDto": {
				"type": "object",
				"properties": {
					"voucherCode": {
						"type": "string"
					}
				},
				"required": [
					"voucherCode"
				]
			},
...
		}
	}
}